+++ apps/dashboard/src/app/(dashboard)/job-listing/[id]/page.tsx
