+++ apps/dashboard/src/app/(dashboard)/job-listing/[id]/provider.tsx
