+++ apps/dashboard/src/app/ModalProvider.tsx
@@ -4,7 +4,7 @@ import { ModalsProvider } from "@mantine/modals"
 import { FC, PropsWithChildren } from "react"
 import { CreateCompanyModal } from "src/modules/company/modals/create-company-modal"
 import { CreateEventModal } from "src/modules/event/modals/create-event-modal"
-import { CreateJobListingModal } from "src/modules/joblisting/modals/create-joblisting-modal"
+import { CreateJobListingModal } from "src/modules/job-listing/modals/create-job-listing-modal"
 
 const modals = {
   "event/create": CreateEventModal,
