+++ apps/db-migrator/src/fixture.ts
@@ -6,7 +6,7 @@ import { eventCommittees } from "./fixtures/committee-organizer"
 import { companies } from "./fixtures/company"
 import { events } from "./fixtures/event"
 import { eventCompany } from "./fixtures/event-company"
-import { jobListingLocationLinks, jobListingLocations, jobListings } from "./fixtures/joblisting"
+import { jobListingLocationLinks, jobListingLocations, jobListings } from "./fixtures/job-listing"
 import { marks } from "./fixtures/mark"
 import { personalMarks } from "./fixtures/personal-mark"
 import { products } from "./fixtures/product"
