+++ packages/gateway-trpc/src/router.ts
@@ -6,7 +6,7 @@ import { paymentRouter } from "./modules/payment/payment-router"
 import { t } from "./trpc"
 import { userRouter } from "./modules/user/user-router"
 import { personalMarkRouter } from "./modules/mark/personal-mark-router"
-import { jobListingRouter } from "./modules/joblisting/joblisting-router"
+import { jobListingRouter } from "./modules/job-listing/job-listing-router"
 
 export const appRouter = t.router({
   committee: committeeRouter,
