+++ packages/types/src/index.ts
@@ -8,4 +8,4 @@ export * from "./payment"
 export * from "./privacy-permissions"
 export * from "./notification-permissions"
 export * from "./event-committee"
-export * from "./joblisting"
+export * from "./job-listing"
