+++ packages/types/src/job-listing.ts
